Bytecode for generated constructor accessors, enclosing classes/methods, ResourceBundle.getClassContext

+++ edu.ubc.mirrors.eclipse.mat/src/edu/ubc/mirrors/eclipse/mat/HeapDumpPrimitiveArrayMirror.java
@@ -72,7 +72,7 @@ public class HeapDumpPrimitiveArrayMirror extends BoxingArrayMirror implements H
     
     @Override
     public double getDouble(int index) throws ArrayIndexOutOfBoundsException {
-        return ((long[])getValueArray())[index];
+        return ((double[])getValueArray())[index];
     }
     
     @Override
